Among Us v2022.03.29 Ron's issue has not been fixed

+++ src/Patches/PassiveButton_ReceiveClickDown.cs
@@ -11,7 +11,7 @@ namespace HarryPotter.Patches
             if (__instance == null || __instance.name == null) return;
 
             if (__instance.name == "SnitchButton" && Main.Instance.GetLocalModdedPlayer().HasItem(3))
-                Main.Instance.RpcForceAllVotes(__instance.transform.GetComponentInParent<PlayerVoteArea>().TargetPlayerId);
+                Main.Instance.RpcForceAllVotes((sbyte)__instance.transform.GetComponentInParent<PlayerVoteArea>().TargetPlayerId);
                 
             if (__instance.name == "SortButton" && Main.Instance.GetLocalModdedPlayer().HasItem(8))
                 Main.Instance.RpcRevealRole((byte)__instance.transform.GetComponentInParent<PlayerVoteArea>().TargetPlayerId);
